fix bug with error

+++ src/examples/Exp1WordCountMapper.java
@@ -2,6 +2,7 @@ package examples;
 
 import io.Context;
 import io.IntWritable;
+import io.LongWritable;
 import io.Text;
 import mapreduce.Mapper;
 
@@ -12,15 +13,15 @@ import mapreduce.Mapper;
  */
 public class Exp1WordCountMapper extends Mapper {
 
-	@Override
-	public void map(IntWritable key, Text value, Context context) {
-		String line = value.get();
-		String[] words = line.split("\\s+");
+    @Override
+    public void map(LongWritable key, Text value, Context context) {
+        String line = value.get();
+        String[] words = line.split("\\s+");
 
-		IntWritable fixVal = new IntWritable(1);
-		for (String word : words) {
-			context.write(new Text(word), fixVal);
-		}
-	}
+        IntWritable fixVal = new IntWritable(1);
+        for (String word : words) {
+            context.write(new Text(word), fixVal);
+        }
+    }
 
 }
